+++ Manatee.Trello/ReadOnlyBoardCollection.cs
@@ -11,7 +11,7 @@ using Manatee.Trello.Json;
 namespace Manatee.Trello
 {
 	/// <summary>
-	/// A read-only collectin of boards.
+	/// A read-only collection of boards.
 	/// </summary>
 	public class ReadOnlyBoardCollection : ReadOnlyCollection<IBoard>, IReadOnlyBoardCollection
 	{